requirements.txt
pandas==0.23.1
PyVimeo==1.0.5
seaborn==0.8.1
matplotlib==2.2.2
scikit-learn==0.19.2
sklearn==0.0
pycurl==7.43.0.2
googletrans==2.3.0